Rename watched-shows to continue-watching
