--- data/watched-shows/api/src/commonMain/kotlin/com/thomaskioko/tvmaniac/watchedshows/api/WatchedShowEntry.kt
@@ -1,9 +1,0 @@
-package com.thomaskioko.tvmaniac.watchedshows.api
-
-public data class WatchedShowEntry(
-    val traktId: Long,
-    val tmdbId: Long?,
-    val plays: Long,
-    val lastWatchedAt: Long,
-    val lastUpdatedAt: Long,
-)
