--- data/watched-shows/api/src/commonMain/kotlin/com/thomaskioko/tvmaniac/watchedshows/api/WatchedShowsDao.kt
@@ -1,18 +1,0 @@
-package com.thomaskioko.tvmaniac.watchedshows.api
-
-import kotlinx.coroutines.flow.Flow
-
-public interface WatchedShowsDao {
-
-    public fun entries(): List<WatchedShowEntry>
-
-    public fun entriesObservable(): Flow<List<WatchedShowEntry>>
-
-    public fun traktIdsMissingShowDetails(): List<Long>
-
-    public fun upsert(entry: WatchedShowEntry)
-
-    public fun deleteByTraktId(traktId: Long)
-
-    public fun deleteAll()
-}
