--- data/watched-shows/api/src/commonMain/kotlin/com/thomaskioko/tvmaniac/watchedshows/api/WatchedShowsRepository.kt
@@ -1,6 +1,0 @@
-package com.thomaskioko.tvmaniac.watchedshows.api
-
-public interface WatchedShowsRepository {
-
-    public suspend fun syncWatchedShows(forceRefresh: Boolean = false)
-}
