--- data/watched-shows/testing/build.gradle.kts
@@ -1,14 +1,0 @@
-plugins {
-    alias(libs.plugins.app.kmp)
-}
-
-kotlin {
-    sourceSets {
-        commonMain {
-            dependencies {
-                api(projects.data.watchedShows.api)
-                api(libs.coroutines.core)
-            }
-        }
-    }
-}
