+++ settings.gradle.kts
@@ -208,9 +208,9 @@ include(
     ":data:user:api",
     ":data:user:implementation",
     ":data:user:testing",
-    ":data:watched-shows:api",
-    ":data:watched-shows:implementation",
-    ":data:watched-shows:testing",
+    ":data:continue-watching:api",
+    ":data:continue-watching:implementation",
+    ":data:continue-watching:testing",
     ":data:watchlist:api",
     ":data:watchlist:implementation",
     ":data:watchlist:testing",
